# docs/api/openapi-v1.yaml
openapi: 3.0.3
info:
  title: LighterPack API
  version: 1.0.0
paths:
  /api/v1/session:
    get:
      summary: Get authenticated session payload
      responses:
        '200':
          description: Session payload
  /api/v1/library:
    put:
      summary: Save an entire library document
      requestBody:
        required: true
      responses:
        '200':
          description: Save success
        '409':
          description: Stale sync token
  /api/v1/packs:
    get:
      summary: List packs for current user
      responses:
        '200':
          description: Pack list
  /api/v1/packs/{packId}/share:
    post:
      summary: Ensure pack has a share external ID
      parameters:
        - in: path
          name: packId
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Created share id

  /api/v1/trips:
    get:
      summary: List trips for current user
      responses:
        '200':
          description: Trip list
    post:
      summary: Create a trip owned by the current user
      responses:
        '201':
          description: Trip created
  /api/v1/trips/{tripId}:
    get:
      summary: Load a single trip including shared members and group gear panel data
      parameters:
        - in: path
          name: tripId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Trip detail
  /api/v1/trips/{tripId}/invitations:
    post:
      summary: Invite a user email as an editor/viewer on a trip
      responses:
        '201':
          description: Invitation created
  /api/v1/trips/{tripId}/accept:
    post:
      summary: Accept invitation and select list visibility to share with trip
      responses:
        '200':
          description: Invitation accepted
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              nullable: true